requirements.txt
actionlib==1.14.0
aiohttp==3.9.3
aiosignal==1.3.1
angles==1.9.13
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bondpy==1.8.6
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2024.2.2
charset-normalizer==3.3.2
copyleaks==4.0.1
cv-bridge==1.16.2
dataclasses-json==0.6.4
diagnostic-updater==1.11.0
distro==1.9.0
frozenlist==1.4.1
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
image-geometry==1.16.2
interactive-markers==1.12.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.10
langchain-community==0.0.25
langchain-core==0.1.28
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langsmith==0.1.14
marshmallow==3.21.0
message-filters==1.16.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.13.3
orjson==3.9.15
packaging==23.2
pillow==10.2.0
pydantic==2.6.3
pydantic_core==2.16.3
pypdf==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-qt-binding==0.4.4
pytz==2024.1
PyYAML==6.0.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-py-common==0.4.2
regex==2023.12.25
requests==2.31.0
resource_retriever==1.12.7
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosparam==1.16.0
rospy==1.16.0
rostest==1.16.0
rosunit==1.15.8
rqt-runtime-monitor==0.5.10
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_py_console==0.4.10
rqt_shell==0.4.11
rqt_top==0.4.10
rqt_web==0.4.10
sensor-msgs==1.13.1
six==1.16.0
smach==2.5.2
smclib==1.8.6
sniffio==1.3.1
SQLAlchemy==2.0.27
tenacity==8.2.3
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
tf2-sensor-msgs==0.7.7
tiktoken==0.6.0
topic-tools==1.16.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
xacro==1.14.17
yarl==1.9.4
